+++ contrib/runners/orquesta_runner/tests/unit/test_error_handling.py
@@ -159,9 +159,11 @@ class OrquestaErrorHandlingTest(st2tests.DbTestCase):
     def test_fail_input_rendering(self):
         expected_errors = [
             {
+                'type': 'error',
                 'message': (
-                    'Unable to evaluate expression \'<% abs(4).value %>\'. '
-                    'NoFunctionRegisteredException: Unknown function "#property#value"'
+                    'YaqlEvaluationException: Unable to evaluate expression '
+                    '\'<% abs(4).value %>\'. NoFunctionRegisteredException: '
+                    'Unknown function "#property#value"'
                 )
             }
         ]
@@ -191,9 +193,11 @@ class OrquestaErrorHandlingTest(st2tests.DbTestCase):
     def test_fail_vars_rendering(self):
         expected_errors = [
             {
+                'type': 'error',
                 'message': (
-                    'Unable to evaluate expression \'<% abs(4).value %>\'. '
-                    'NoFunctionRegisteredException: Unknown function "#property#value"'
+                    'YaqlEvaluationException: Unable to evaluate expression '
+                    '\'<% abs(4).value %>\'. NoFunctionRegisteredException: '
+                    'Unknown function "#property#value"'
                 )
             }
         ]
@@ -223,9 +227,11 @@ class OrquestaErrorHandlingTest(st2tests.DbTestCase):
     def test_fail_start_task_action(self):
         expected_errors = [
             {
+                'type': 'error',
                 'message': (
-                    'Unable to evaluate expression \'<% ctx().func.value %>\'. '
-                    'NoFunctionRegisteredException: Unknown function "#property#value"'
+                    'YaqlEvaluationException: Unable to evaluate expression '
+                    '\'<% ctx().func.value %>\'. NoFunctionRegisteredException: '
+                    'Unknown function "#property#value"'
                 ),
                 'task_id': 'task1'
             }
@@ -256,9 +262,11 @@ class OrquestaErrorHandlingTest(st2tests.DbTestCase):
     def test_fail_start_task_input_expr_eval(self):
         expected_errors = [
             {
+                'type': 'error',
                 'message': (
-                    'Unable to evaluate expression \'<% ctx().msg1.value %>\'. '
-                    'NoFunctionRegisteredException: Unknown function "#property#value"'
+                    'YaqlEvaluationException: Unable to evaluate expression '
+                    '\'<% ctx().msg1.value %>\'. NoFunctionRegisteredException: '
+                    'Unknown function "#property#value"'
                 ),
                 'task_id': 'task1'
             }
@@ -293,8 +301,11 @@ class OrquestaErrorHandlingTest(st2tests.DbTestCase):
         else:
             msg = 'Value "{u\'x\': u\'foobar\'}" must either be a string or None. Got "dict".'
 
+        msg = 'ValueError: ' + msg
+
         expected_errors = [
             {
+                'type': 'error',
                 'message': msg,
                 'task_id': 'task1'
             }
@@ -328,9 +339,11 @@ class OrquestaErrorHandlingTest(st2tests.DbTestCase):
     def test_fail_next_task_action(self):
         expected_errors = [
             {
+                'type': 'error',
                 'message': (
-                    'Unable to evaluate expression \'<% ctx().func.value %>\'. '
-                    'NoFunctionRegisteredException: Unknown function "#property#value"'
+                    'YaqlEvaluationException: Unable to evaluate expression '
+                    '\'<% ctx().func.value %>\'. NoFunctionRegisteredException: '
+                    'Unknown function "#property#value"'
                 ),
                 'task_id': 'task2'
             }
@@ -371,9 +384,11 @@ class OrquestaErrorHandlingTest(st2tests.DbTestCase):
     def test_fail_next_task_input_expr_eval(self):
         expected_errors = [
             {
+                'type': 'error',
                 'message': (
-                    'Unable to evaluate expression \'<% ctx().msg2.value %>\'. '
-                    'NoFunctionRegisteredException: Unknown function "#property#value"'
+                    'YaqlEvaluationException: Unable to evaluate expression '
+                    '\'<% ctx().msg2.value %>\'. NoFunctionRegisteredException: '
+                    'Unknown function "#property#value"'
                 ),
                 'task_id': 'task2'
             }
@@ -417,8 +432,11 @@ class OrquestaErrorHandlingTest(st2tests.DbTestCase):
         else:
             msg = 'Value "{u\'x\': u\'foobar\'}" must either be a string or None. Got "dict".'
 
+        msg = 'ValueError: ' + msg
+
         expected_errors = [
             {
+                'type': 'error',
                 'message': msg,
                 'task_id': 'task2'
             }
@@ -463,6 +481,7 @@ class OrquestaErrorHandlingTest(st2tests.DbTestCase):
     def test_fail_task_execution(self):
         expected_errors = [
             {
+                'type': 'error',
                 'message': 'Execution failed. See result for details.',
                 'task_id': 'task1',
                 'result': {
@@ -505,8 +524,9 @@ class OrquestaErrorHandlingTest(st2tests.DbTestCase):
     def test_fail_task_transition(self):
         expected_errors = [
             {
+                'type': 'error',
                 'message': (
-                    "Unable to resolve key 'foobar' in expression "
+                    "YaqlEvaluationException: Unable to resolve key 'foobar' in expression "
                     "'<% succeeded() and result().foobar %>' from context."
                 ),
                 'task_transition_id': 'task2__0',
@@ -548,9 +568,11 @@ class OrquestaErrorHandlingTest(st2tests.DbTestCase):
     def test_fail_task_publish(self):
         expected_errors = [
             {
+                'type': 'error',
                 'message': (
-                    'Unable to evaluate expression \'<% foobar() %>\'. '
-                    'NoFunctionRegisteredException: Unknown function "foobar"'
+                    'YaqlEvaluationException: Unable to evaluate expression '
+                    '\'<% foobar() %>\'. NoFunctionRegisteredException: '
+                    'Unknown function "foobar"'
                 ),
                 'task_transition_id': 'task2__0',
                 'task_id': 'task1'
@@ -591,9 +613,11 @@ class OrquestaErrorHandlingTest(st2tests.DbTestCase):
     def test_fail_output_rendering(self):
         expected_errors = [
             {
+                'type': 'error',
                 'message': (
-                    'Unable to evaluate expression \'<% abs(4).value %>\'. '
-                    'NoFunctionRegisteredException: Unknown function "#property#value"'
+                    'YaqlEvaluationException: Unable to evaluate expression '
+                    '\'<% abs(4).value %>\'. NoFunctionRegisteredException: '
+                    'Unknown function "#property#value"'
                 )
             }
         ]
@@ -620,6 +644,67 @@ class OrquestaErrorHandlingTest(st2tests.DbTestCase):
         wf_ex_db = wf_db_access.WorkflowExecution.get_by_id(wf_ex_db.id)
         self.assertEqual(wf_ex_db.status, wf_states.FAILED)
         self.assertListEqual(self.sort_wf_runtime_errors(wf_ex_db.errors), expected_errors)
+
+        lv_ac_db = lv_db_access.LiveAction.get_by_id(str(lv_ac_db.id))
+        self.assertEqual(lv_ac_db.status, ac_const.LIVEACTION_STATUS_FAILED)
+        self.assertDictEqual(lv_ac_db.result, expected_result)
+
+        ac_ex_db = ex_db_access.ActionExecution.get_by_id(str(ac_ex_db.id))
+        self.assertEqual(ac_ex_db.status, ac_const.LIVEACTION_STATUS_FAILED)
+        self.assertDictEqual(ac_ex_db.result, expected_result)
+
+    def test_output_on_error(self):
+        expected_output = {
+            'progress': 25
+        }
+
+        expected_errors = [
+            {
+                'type': 'error',
+                'task_id': 'task2',
+                'message': 'Execution failed. See result for details.',
+                'result': {
+                    'failed': True,
+                    'return_code': 1,
+                    'stderr': '',
+                    'stdout': '',
+                    'succeeded': False
+                }
+            }
+        ]
+
+        expected_result = {
+            'errors': expected_errors,
+            'output': expected_output
+        }
+
+        wf_meta = base.get_wf_fixture_meta_data(TEST_PACK_PATH, 'output-on-error.yaml')
+        lv_ac_db = lv_db_models.LiveActionDB(action=wf_meta['name'])
+        lv_ac_db, ac_ex_db = ac_svc.request(lv_ac_db)
+        wf_ex_db = wf_db_access.WorkflowExecution.query(action_execution=str(ac_ex_db.id))[0]
+
+        # Assert task1 is already completed and workflow execution is still running.
+        query_filters = {'workflow_execution': str(wf_ex_db.id), 'task_id': 'task1'}
+        tk1_ex_db = wf_db_access.TaskExecution.query(**query_filters)[0]
+        tk1_ac_ex_db = ex_db_access.ActionExecution.query(task_execution=str(tk1_ex_db.id))[0]
+        tk1_lv_ac_db = lv_db_access.LiveAction.get_by_id(tk1_ac_ex_db.liveaction['id'])
+        self.assertEqual(tk1_lv_ac_db.status, ac_const.LIVEACTION_STATUS_SUCCEEDED)
+        wf_svc.handle_action_execution_completion(tk1_ac_ex_db)
+        wf_ex_db = wf_db_access.WorkflowExecution.get_by_id(wf_ex_db.id)
+        self.assertEqual(wf_ex_db.status, wf_states.RUNNING)
+
+        # Assert task2 is already completed and workflow execution has failed.
+        query_filters = {'workflow_execution': str(wf_ex_db.id), 'task_id': 'task2'}
+        tk2_ex_db = wf_db_access.TaskExecution.query(**query_filters)[0]
+        tk2_ac_ex_db = ex_db_access.ActionExecution.query(task_execution=str(tk2_ex_db.id))[0]
+        tk2_lv_ac_db = lv_db_access.LiveAction.get_by_id(tk2_ac_ex_db.liveaction['id'])
+        self.assertEqual(tk2_lv_ac_db.status, ac_const.LIVEACTION_STATUS_FAILED)
+        wf_svc.handle_action_execution_completion(tk2_ac_ex_db)
+
+        # Check output and result for expected value(s).
+        wf_ex_db = wf_db_access.WorkflowExecution.get_by_id(wf_ex_db.id)
+        self.assertEqual(wf_ex_db.status, wf_states.FAILED)
+        self.assertDictEqual(wf_ex_db.output, expected_output)
 
         lv_ac_db = lv_db_access.LiveAction.get_by_id(str(lv_ac_db.id))
         self.assertEqual(lv_ac_db.status, ac_const.LIVEACTION_STATUS_FAILED)