+++ contrib/runners/orquesta_runner/tests/unit/test_functions_task.py
@@ -196,9 +196,11 @@ class OrquestaFunctionTest(st2tests.DbTestCase):
 
         expected_errors = [
             {
+                'type': 'error',
                 'message': (
-                    'Unable to evaluate expression \'<% task("task0") %>\'. '
-                    'ExpressionEvaluationException: Unable to find task execution for "task0".'
+                    'YaqlEvaluationException: Unable to evaluate expression '
+                    '\'<% task("task0") %>\'. ExpressionEvaluationException: '
+                    'Unable to find task execution for "task0".'
                 ),
                 'task_transition_id': 'noop__0',
                 'task_id': 'task1'
@@ -222,9 +224,11 @@ class OrquestaFunctionTest(st2tests.DbTestCase):
 
         expected_errors = [
             {
+                'type': 'error',
                 'message': (
-                    'Unable to evaluate expression \'{{ task("task0") }}\'. '
-                    'ExpressionEvaluationException: Unable to find task execution for "task0".'
+                    'JinjaEvaluationException: Unable to evaluate expression '
+                    '\'{{ task("task0") }}\'. ExpressionEvaluationException: '
+                    'Unable to find task execution for "task0".'
                 ),
                 'task_transition_id': 'noop__0',
                 'task_id': 'task1'